provinces/pyproject.toml
[project]
name = "vn-open-api-provinces"
description = ""
authors = [{ name = "Nguyễn Hồng Quân", email = "ng.hong.quan@gmail.com" }]
license = { text = "GPL-3.0-or-later" }
requires-python = "<4.0,>=3.12"
dependencies = [
  "aiofiles >= 24.1.0",
  "Unidecode >= 1.3.8",
  "vietnam-provinces @ git+https://github.com/sunshine-tech/VietnamProvinces.git@4546c0c",
  "logbook >= 1.8.1",
  "fastapi >= 0.116.1",
  "pydantic >= 2.11.1",
  "lunr >= 0.8.0",
  "pydantic-settings >= 2.10.1",
  "fastenumplus>=1.4.0",
  "granian[reload]>=2.5.0",
  "fastapi-problem>=0.11.4",
  "uvicorn>=0.35.0",
  "python-dotenv>=1.1.1",
]
dynamic = ["version"]

[tool.pdm.version]
source = 'file'
path = "api/__init__.py"

[tool.ruff]
line-length = 120
# Exclude a variety of commonly ignored directories.
exclude = [
  ".bzr",
  ".direnv",
  ".eggs",
  ".git",
  ".hg",
  ".mypy_cache",
  ".nox",
  ".pants.d",
  ".ruff_cache",
  ".svn",
  ".tox",
  ".venv",
  "__pypackages__",
  "_build",
  "buck-out",
  "build",
  "dist",
  "node_modules",
  "venv",
]
target-version = "py312"

[tool.ruff.lint]
# Enable Pyflakes `E` and `F` codes by default.
select = ['E', 'F', 'I']
ignore = []
# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "single"

[tool.ruff.lint.isort]
lines-after-imports = 2

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[dependency-groups]
dev = [
    "mypy >= 1.17.0",
    "pytest >= 8.3.2",
    "ruff >= 0.12.5",
]
lint = [
    "ruff>=0.12.5",
]

[tool.mypy]
python_version = "3.12"
show_error_codes = true

[[tool.mypy.overrides]]
module = [
  "logbook.*",
  "lunr.*",
]
ignore_missing_imports = true
